Split config in multiple files
