+++ package.json
@@ -21,8 +21,8 @@
     }
   ],
   "scripts": {
-    "start": "NODE_ENV=\"development\" webpack-dev-server --colors",
-    "starts": "NODE_ENV=\"development\" webpack-dev-server --https --colors",
+    "start": "NODE_ENV=\"development\" webpack-dev-server --colors --config ./build_scripts/webpack.config.js",
+    "starts": "NODE_ENV=\"development\" webpack-dev-server --https --colors --config ./build_scripts/webpack.config.js",
     "precommit": "lint-staged",
     "format": "prettier-eslint 'src/**/!(*.min).js' 'src/**/*.jsx' --write",
     "lint": "eslint --fix --ext .js --ext .jsx ./",
@@ -31,7 +31,7 @@
     "e2e-local": "CYPRESS_E2E_TEST_ENV=\"local\" cypress run",
     "e2e-local-open": "CYPRESS_E2E_TEST_ENV=\"local\" cypress open",
     "dev": "jest --watch",
-    "build": "NODE_ENV=\"production\" webpack",
+    "build": "NODE_ENV=\"production\" webpack --config ./build_scripts/webpack.config.js",
     "prepare-jar": "node ./scripts/prepare-mvn-package.js",
     "jar": "yarn build && mvn package",
     "version-pom": "node ./scripts/set-pom-version.js -f ./pom.xml -v $npm_package_version",
